Fixed Image size and ratio issues 2
We snapshotted the entire canvas even though the opened image doesn't fill all of its space. This would cause a blank canvas area to written to the new buffered image.

The fix was to simply take the new size of the image after passing it through image constructor , which holds the actual image dimensions. Next, pass it through as args to the new Writable Image constructor,  so we can be returned a writable image obj with the correct dimensions.